Fixing CI build issues for Focal noetic.
The error might have stemmed from not specifying the
distribution of interest to either Travis and/or to
rosdep.

Additionally, PYTHONPATH had to be defined as a
distribution specific environment variable, mainly
because ROS Noetic migrated from Python 2 to Python
3.

Finally, Travis CI's default builds with Ubuntu 20.04
uses an 'older' version of GCC that exhibits a
bug while compiling. This required an `apt upgrade`
on GCC.

Note that, as is, the build will still report an
error. That is because `grasp_synergy` - an external
dependency for one of the projects - requires
installation of `python-sklearn`, which is not a listed
Debian package in Ubuntu 20.04 (Python 2 -> 3
transition on Ubuntu's end too...). `grasp_synergy`
could be altered to a different dependency supporting
ROS Noetic to see that CI builds successfully.
I had success with `ros-teleop/teleop_twist_keyboard`.

Signed-off-by: aprotyas <aprotyas@u.rochester.edu>